Split karva_metadata::options into submodules
